Make an exploration with all prototype interactions work end-to-end.
This includes:
- Introducing an exploration with all prototype interactions
- Introducing support for binding all interaction types
- Fixing some parsing issues with fractions
- Rewiring the interaction object creation for views to be generic
- Removing temporary answer handling everywhere
- Other random fixes

This includes no tests, and there are a few other fixes still necessary
for the core learning experience to fully work.

+++ domain/src/main/java/org/oppia/domain/classify/rules/numberwithunits/NumberWithUnitsIsEqualToRuleClassifierProvider.kt
@@ -42,7 +42,7 @@ internal class NumberWithUnitsIsEqualToRuleClassifierProvider @Inject constructo
     }
   }
 
-  private fun realMatches(answer: Float, input: Float): Boolean {
+  private fun realMatches(answer: Double, input: Double): Boolean {
     return input.approximatelyEquals(answer)
   }
